requirements.txt
# ================================
# InferenceHub - Dependencies
# ================================

# Web Framework
fastapi==0.109.0
uvicorn==0.27.0

# Database
sqlalchemy==2.0.25
psycopg2-binary==2.9.9

# AI
groq==0.11.0
httpx==0.27.0

# Utilities
python-dotenv==1.0.0
